requirements.txt
selenium
webdriver-manager
mysql-connector-python
python-dotenv